+++ lib/ccl_ui.dart
@@ -1,4 +1,3 @@
 library ccl_ui;
 
-export 'common/common.dart';
-export 'widgets/widgets.dart';
+export 'src/ui.dart';