+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 
-<a name="v0.2.5"></a>
-## [v0.2.5](https://github.com/CeylonCodeLabs/ccl_packages/compare/v0.2.4...v0.2.5) (2024-06-26)
+<a name="v0.6.0"></a>
+## [v0.6.0](https://github.com/CeylonCodeLabs/ccl_packages/compare/v0.5.0...v0.6.0) (2024-09-11)
 
 ### Chore
 
-* **Release:** v0.2.5
+* **Release:** v0.6.0
 
 ### Docs
 
@@ -12,7 +12,89 @@
 
 ### Refactor
 
+* **Code:** Re-arranged
+
+
+<a name="v0.5.0"></a>
+## [v0.5.0](https://github.com/CeylonCodeLabs/ccl_packages/compare/v0.4.0...v0.5.0) (2024-09-05)
+
+### Build
+
+* **Package:** Package update changes
+
+### Chore
+
+* **Release:** v0.5.0
+
+### Docs
+
+* **Release:** Updates
+
+
+<a name="v0.4.0"></a>
+## [v0.4.0](https://github.com/CeylonCodeLabs/ccl_packages/compare/v0.3.0...v0.4.0) (2024-09-05)
+
+### Build
+
+* **Package:** Package update changes
+
+### Chore
+
+* **Release:** v0.4.0
+
+### Docs
+
+* **Release:** Updates
+
+
+<a name="v0.3.0"></a>
+## [v0.3.0](https://github.com/CeylonCodeLabs/ccl_packages/compare/v0.2.6...v0.3.0) (2024-08-18)
+
+### Chore
+
+* **Release:** v0.3.0
+
+### Docs
+
+* **Release:** Updates
+
+### Feat
+
+* **Background Progress:** Added progressIndicator and isChildVisibleWhileBusy params
+
+### Refactor
+
+* **Code:** Arrangement optimization
+
+### Test
+
+* **Background Progress:** Add tests for the widget
+
+
+<a name="v0.2.6"></a>
+## [v0.2.6](https://github.com/CeylonCodeLabs/ccl_packages/compare/v0.2.4...v0.2.6) (2024-08-16)
+
+### Chore
+
+* **Release:** v0.2.6
+* **Release:** v0.2.5
+* **Release:** v0.2.5
+
+### Docs
+
+* **Release:** Change log and release updates
+* **Release:** Updates
+* **Release:** Updates
+
+### Refactor
+
+* **Code:** Export required features
 * **Localization:** Implementation updates
+
+### Reverts
+
+* chore(Release): v0.2.5
+* docs(Release): Change log and release updates
 
 
 <a name="v0.2.4"></a>